restructured components into folders
