+++ src/webparts/v2NpmBanner/components/PropPaneCols/components/fetch/funcions.ts
@@ -1,7 +1,7 @@
 
 import { IWeb, Web, } from "@pnp/sp/presets/all";
-import { IMinField, IMinListProps, IsEditable, IFieldPanelFetchState } from "./IPropPaneColsProps";
-import { getHelpfullErrorV2 } from '../../../fpsReferences';
+import { IMinField, IMinListProps, IsEditable, IFieldPanelFetchState } from "../IPropPaneColsProps";
+import { getHelpfullErrorV2 } from '../../../../fpsReferences';
 
   // export async function clickFetchFields(  list: IMinListProps, setState: any, updatePerformance: any ) : Promise<void> {
 export async function fetchFields(  list: IMinListProps ) : Promise<IFieldPanelFetchState> {
