+++ src/webparts/v2NpmBanner/components/PropPaneCols/components/main/FieldTable.tsx
@@ -3,7 +3,7 @@ import { useState, useEffect } from 'react';
 import { Panel, PanelType } from 'office-ui-fabric-react/lib/Panel';
 import ReactJson from 'react-json-view';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { getHighlightedText , getHelpfullErrorV2 } from '../../../fpsReferences';
+import { getHighlightedText , getHelpfullErrorV2 } from '../../../../fpsReferences';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { IGrouping, IViewField } from "@pnp/spfx-controls-react/lib/ListView";
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -15,8 +15,8 @@ import { Icon, } from 'office-ui-fabric-react/lib/Icon';
 
 // import styles from '../PropPaneCols.module.scss';
 
-import { IMinField } from "./IPropPaneColsProps";
-import SelectedItemPanelHook from "./FieldPanel";
+import { IMinField } from "../IPropPaneColsProps";
+import SelectedItemPanelHook from "../FieldPanel";
 
 export interface IMainFieldTableHookProps {
   // selected: IMinField[];
