+++ src/webparts/v2NpmBanner/components/PropPaneCols/components/views/Accordion.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { getHighlightedText , getHelpfullErrorV2 } from '../../../fpsReferences';
+import { getHighlightedText , getHelpfullErrorV2 } from '../../../../fpsReferences';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { IGrouping, IViewField } from "@pnp/spfx-controls-react/lib/ListView";
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -13,8 +13,8 @@ import { Icon, } from 'office-ui-fabric-react/lib/Icon';
 
 // import styles from '../PropPaneCols.module.scss';
 
-import { IMinField } from "./IPropPaneColsProps";
-import { createThisViewField } from './ViewFields';
+import { IMinField } from "../IPropPaneColsProps";
+import { createThisViewField } from './functions';
 import ReactJson from 'react-json-view';
 import Accordion from '@mikezimm/npmfunctions/dist/zComponents/Accordion/Accordion';
